+++ app/javascript/css/browser_tab.css.scss
@@ -1,4 +1,3 @@
-
 .tab-browser {
   background: white;
   width: 213px;
@@ -6,10 +5,8 @@
   overflow: hidden;
   height: 100%;
   padding: 1px 0;
-
   border-radius: 3px;
   box-shadow: 2px -2px 2px -2px black, -2px -2px 2px -2px black;
-
   padding-left: 8px;
   display: block;
 }
@@ -23,7 +20,6 @@
   vertical-align: middle;
   width: 16px;
   height: 16px;
-
   margin-right: 7px;
 }
