+++ app/javascript/packs/application.js
@@ -24,7 +24,6 @@ import 'plugins/flatpickr';
 import "@fortawesome/fontawesome-free/css/all"
 
 import "../js/calculators/edit"
-import "../js/account/read_url"
 import "../js/account/handle_browser_tab"
 import "../ajax/calculate_result_button"
