+++ app/javascript/stylesheets/application.scss
@@ -12,7 +12,7 @@ $md: 1024px;
 @import "../css/message.css.scss";
 @import "../css/account_edit_show.css.scss";
 @import "../css/calculator.css.scss";
-@import "../css/browser_bar.css.scss";
+@import "../css/browser_tab.css.scss";
 
 @font-face {
   font-family: 'Comfortaa';
